moved run out of main
